fix(auth): address deep-dive review findings for PR #33
1. Extract _OtpVerificationView and _PhoneFormView from register_screen.dart
   (323 lines, max 200) into separate widget files under presentation/widgets/
2. Align DeelmarktAnimation.standard to 200ms per tokens.md (was 300ms)
3. Replace hardcoded BorderRadius.circular(2) with DeelmarktRadius.xxs token
   and hardcoded Duration(milliseconds: 200) with DeelmarktAnimation.standard
   in password_strength_indicator.dart
4. Derive link TextStyles from theme.textTheme.bodyMedium?.copyWith() instead
   of bare const TextStyle() in registration_form.dart
