+++ lib/core/design_system/radius.dart
@@ -3,6 +3,7 @@
 class DeelmarktRadius {
   DeelmarktRadius._();
 
+  static const double xxs = 2;
   static const double xs = 6;
   static const double sm = 8;
   static const double md = 10;
